+++ fe/src/components/Budget/BudgetView.tsx
@@ -4,6 +4,7 @@ import { useMemo, useState } from 'react';
 import AnalysisEmpty from '@/components/analysis/common/AnalysisEmpty';
 import AnalysisLoading from '@/components/analysis/common/AnalysisLoading';
 import AnalysisSection from '@/components/analysis/common/AnalysisSection';
+import BudgetOverview from './BudgetOverview';
 import BudgetRecommendDialog from '@/components/budget/BudgetRecommendDialog';
 import BudgetSetupDialog from '@/components/budget/BudgetSetupDialog';
 import { Button } from '@/components/ui/button';
@@ -74,14 +75,6 @@ const BudgetView = ({ selectedDate }: { selectedDate: Date }) => {
     })
     .sort((a, b) => b.amount - a.amount);
 
-  const budgetAmount = totalBudget?.amount || 0;
-  const remaining = budgetAmount - totalExpense - futureFixedAmount; // 고정비까지 고려
-  // (실제 지출 + 지출 예정 합산) 퍼센트
-  const totalPercentage = Math.min(
-    Math.round(((totalExpense + futureFixedAmount) / budgetAmount) * 100),
-    100
-  );
-
   const openCategorySetting = (key: string) => {
     setActiveCategoryKey(key);
     setIsCategoryPanelOpen(true);
@@ -150,131 +143,16 @@ const BudgetView = ({ selectedDate }: { selectedDate: Date }) => {
           </AnalysisEmpty>
         </div>
       ) : (
+        // 예산이 있을 때 보여줄 화면
         <>
-          <div className="relative">
-            {/* 예산이 있을 때 보여줄 화면  */}
-            <AnalysisSection title={'총 예산'}>
-              <div className="flex flex-col items-center gap-4 py-6">
-                <div className="absolute top-8 right-8 flex gap-1">
-                  <Button
-                    variant="ghost"
-                    className="text-muted-foreground hover:text-foreground h-8 cursor-pointer px-2"
-                    onClick={() => setIsDialogOpen(true)}
-                  >
-                    수정
-                  </Button>
-
-                  <Button
-                    variant="ghost"
-                    className="text-destructive hover:text-destructive h-8 cursor-pointer px-2"
-                    onClick={() => setIsTotalConfirmOpen(true)}
-                  >
-                    초기화
-                  </Button>
-                </div>
-
-                {/* 메인 콘텐츠 */}
-                <div className="text-center">
-                  <p className="text-muted-foreground text-base font-medium">
-                    이번 달 총 예산
-                  </p>
-                  <p className="text-primary text-3xl font-bold tracking-tight">
-                    {totalBudget.amount.toLocaleString()}
-                    <span className="text-foreground text-lg font-normal">
-                      {' '}
-                      원
-                    </span>
-                  </p>
-                </div>
-
-                {/* 예산 사용 현황 */}
-                <div className="flex w-full max-w-lg flex-col items-center justify-center gap-10 py-4 text-center sm:flex-row">
-                  <div className="flex-1">
-                    <p className="text-muted-foreground">현재 지출</p>
-                    <p
-                      className={`text-lg font-semibold ${THEME_COLOR.EXPENSE}`}
-                    >
-                      {totalExpense.toLocaleString()}원
-                    </p>
-                  </div>
-
-                  {futureFixedAmount > 0 && (
-                    <div className="flex-1">
-                      <p className="text-muted-foreground">지출 예정</p>
-                      <p className="text-lg font-semibold text-[#5C7AFF]">
-                        {futureFixedAmount.toLocaleString()}원
-                      </p>
-                    </div>
-                  )}
-
-                  <div className="flex-1">
-                    <p className="text-muted-foreground">남은 예산</p>
-                    <p
-                      className={cn(
-                        'text-lg font-semibold',
-                        remaining < 0 ? THEME_COLOR.EXPENSE : THEME_COLOR.INCOME
-                      )}
-                    >
-                      {remaining.toLocaleString()}원
-                    </p>
-                  </div>
-                </div>
-
-                {/* 바 차트 */}
-                <div className="w-full max-w-lg space-y-2">
-                  <div className="text-muted-foreground flex justify-between text-sm">
-                    <span>
-                      예산 사용률
-                      {futureFixedAmount > 0 && (
-                        <span className="rounded pl-1 text-sm">
-                          (지출 예정 포함)
-                        </span>
-                      )}
-                    </span>
-                    <span
-                      className={cn(
-                        'font-medium',
-                        totalPercentage >= 90
-                          ? THEME_COLOR.EXPENSE
-                          : 'text-foreground'
-                      )}
-                    >
-                      {totalPercentage}%
-                    </span>
-                  </div>
-
-                  <Progress
-                    value={totalPercentage}
-                    className="h-4"
-                    indicatorClassName={
-                      totalPercentage >= 90 ? 'bg-red-400' : 'bg-primary'
-                    }
-                  />
-
-                  {totalPercentage >= 100 && (
-                    <p className="mt-4 text-center text-sm font-medium">
-                      이번 달 예산을{' '}
-                      <span
-                        className={cn('font-semibold', THEME_COLOR.EXPENSE)}
-                      >
-                        {Math.abs(remaining).toLocaleString()}원 초과
-                      </span>
-                      하여 지출하고 있어요!
-                    </p>
-                  )}
-                  {totalPercentage >= 90 && totalPercentage < 100 && (
-                    <p className="mt-4 text-center text-sm font-medium">
-                      이번 달 예산이{' '}
-                      <span className={cn('font-semibold', THEME_COLOR.INCOME)}>
-                        {remaining.toLocaleString()}원
-                      </span>
-                      밖에 남지 않았습니다.
-                    </p>
-                  )}
-                </div>
-              </div>
-            </AnalysisSection>
-          </div>
+          {/* 총 예산 섹션 */}
+          <BudgetOverview
+            totalAmount={totalBudget.amount}
+            totalExpense={totalExpense}
+            futureFixedAmount={futureFixedAmount}
+            onEdit={() => setIsDialogOpen(true)}
+            onReset={() => setIsTotalConfirmOpen(true)}
+          />
 
           <div className="relative">
             {/* 카테고리별 예산 */}
